refactor(core): split search query and row hydration
