+++ src/basic_memory/repository/postgres_search_repository.py
@@ -16,10 +16,10 @@ from basic_memory.config import BasicMemoryConfig, ConfigManager
 from basic_memory.repository.embedding_provider import EmbeddingProvider
 from basic_memory.repository.embedding_provider_factory import create_embedding_provider
 from basic_memory.repository.search_index_row import SearchIndexRow
+from basic_memory.repository.search_query import relaxed_query_words
 from basic_memory.repository.search_repository_base import (
     SearchRepositoryBase,
     VectorChunkState,
-    relaxed_query_words,
 )
 from basic_memory.repository.metadata_filters import parse_metadata_filters
 from basic_memory.repository.semantic_errors import SemanticDependenciesMissingError
@@ -1066,31 +1066,7 @@ class PostgresSearchRepository(SearchRepositoryBase):
             logger.error(f"Database error during search: {e}")
             raise
 
-        results = [
-            SearchIndexRow(
-                project_id=self.project_id,
-                id=row.id,
-                title=row.title,
-                permalink=row.permalink,
-                file_path=row.file_path,
-                type=row.type,
-                score=float(row.score) if row.score else 0.0,
-                metadata=(
-                    row.metadata
-                    if isinstance(row.metadata, dict)
-                    else (json.loads(row.metadata) if row.metadata else {})
-                ),
-                from_id=row.from_id,
-                to_id=row.to_id,
-                relation_type=row.relation_type,
-                entity_id=row.entity_id,
-                content_snippet=row.content_snippet,
-                category=row.category,
-                created_at=row.created_at,
-                updated_at=row.updated_at,
-            )
-            for row in rows
-        ]
+        results = [SearchIndexRow.from_mapping(row._asdict()) for row in rows]
 
         logger.trace(f"Found {len(results)} search results")
         for r in results:
