+++ tests/repository/test_search_index_row.py
@@ -1,8 +1,33 @@
 """Tests for SearchIndexRow data structure."""
 
 from datetime import datetime
+from decimal import Decimal
 
 from basic_memory.repository.search_index_row import SearchIndexRow
+
+
+def test_from_mapping_normalizes_database_values():
+    """Database rows share one hydration path across SQLite and Postgres."""
+    now = datetime.now()
+    row = SearchIndexRow.from_mapping(
+        {
+            "project_id": 1,
+            "id": 2,
+            "type": "observation",
+            "file_path": "notes/example.md",
+            "created_at": now,
+            "updated_at": now,
+            "metadata": '{"tags": ["example"]}',
+            "score": Decimal("0.25"),
+            "entity_id": 3,
+            "content_snippet": "Shared hydration",
+        }
+    )
+
+    assert row.metadata == {"tags": ["example"]}
+    assert row.score == 0.25
+    assert row.entity_id == 3
+    assert row.content_snippet == "Shared hydration"
 
 
 def test_content_display_limit_is_4000():
